# examples/chi-basic/testdata/expected/chi/goduct_routes.go
// Code generated by goduct. DO NOT EDIT.

package api

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/coder/websocket"
	"github.com/go-chi/chi/v5"

	goduct "github.com/townsendmerino/goduct/runtime"
)

// Register wires every goduct-annotated handler in this package to r.
// Call this after applying any middleware you want to share across handlers.
func Register(r chi.Router) {
	r.Get("/users/{id}", handleGetUser)
	r.Get("/users", handleListUsers)
	r.Post("/users", handleCreateUser)
	r.Patch("/users/{id}", handleUpdateUser)
	r.Delete("/users/{id}", handleDeleteUser)
	r.Post("/users/{id}/avatar", handleUploadAvatar)
	r.Get("/users/{id}/events", handleWatchUserEvents)
	r.Get("/users/{id}/echo", handleEcho)
}

func handleGetUser(w http.ResponseWriter, r *http.Request) {
	var req GetUserRequest
	req.ID = chi.URLParam(r, "id")
	q := r.URL.Query()
	req.Include = q.Get("include")
	resp, err := GetUser(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	goduct.WriteJSON(w, http.StatusOK, resp)
}

func handleListUsers(w http.ResponseWriter, r *http.Request) {
	var req ListUsersRequest
	q := r.URL.Query()
	if v := q.Get("limit"); v != "" {
		n, err := strconv.Atoi(v)
		if err != nil {
			goduct.WriteError(w, goduct.BadRequest("limit must be an integer"))
			return
		}
		req.Limit = n
	}
	req.Cursor = q.Get("cursor")
	if v := q.Get("active"); v != "" {
		n, err := strconv.ParseBool(v)
		if err != nil {
			goduct.WriteError(w, goduct.BadRequest("active must be a boolean"))
			return
		}
		req.Active = n
	}
	if v := q.Get("minScore"); v != "" {
		n, err := strconv.ParseFloat(v, 64)
		if err != nil {
			goduct.WriteError(w, goduct.BadRequest("minScore must be a number"))
			return
		}
		req.MinScore = n
	}
	resp, err := ListUsers(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	goduct.WriteJSON(w, http.StatusOK, resp)
}

func handleCreateUser(w http.ResponseWriter, r *http.Request) {
	var req CreateUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		goduct.WriteError(w, goduct.BadRequest("invalid json body"))
		return
	}
	resp, err := CreateUser(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	goduct.WriteJSON(w, http.StatusCreated, resp)
}

func handleUpdateUser(w http.ResponseWriter, r *http.Request) {
	var req UpdateUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		goduct.WriteError(w, goduct.BadRequest("invalid json body"))
		return
	}
	req.ID = chi.URLParam(r, "id")
	resp, err := UpdateUser(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	goduct.WriteJSON(w, http.StatusOK, resp)
}

func handleDeleteUser(w http.ResponseWriter, r *http.Request) {
	var req DeleteUserRequest
	req.ID = chi.URLParam(r, "id")
	if err := DeleteUser(r.Context(), req); err != nil {
		goduct.WriteError(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func handleUploadAvatar(w http.ResponseWriter, r *http.Request) {
	var req UploadAvatarRequest
	if err := r.ParseMultipartForm(32 << 20); err != nil {
		goduct.WriteError(w, goduct.BadRequest("invalid multipart form"))
		return
	}
	if files := r.MultipartForm.File["file"]; len(files) > 0 {
		req.File = files[0]
	} else {
		goduct.WriteError(w, goduct.BadRequest("file is required"))
		return
	}
	if req.File != nil && req.File.Size > 1048576 {
		goduct.WriteError(w, goduct.BadRequest("file exceeds 1048576 byte limit"))
		return
	}
	if files := r.MultipartForm.File["thumbnails"]; len(files) > 0 {
		req.Thumbnails = files
	}
	if vs := r.MultipartForm.Value["caption"]; len(vs) > 0 {
		req.Caption = vs[0]
	}
	req.UserID = chi.URLParam(r, "id")
	resp, err := UploadAvatar(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	goduct.WriteJSON(w, http.StatusCreated, resp)
}

func handleWatchUserEvents(w http.ResponseWriter, r *http.Request) {
	var req WatchUserEventsRequest
	req.ID = chi.URLParam(r, "id")
	ch, err := WatchUserEvents(r.Context(), req)
	if err != nil {
		goduct.WriteError(w, err)
		return
	}
	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Set("Cache-Control", "no-cache")
	w.WriteHeader(http.StatusOK)
	goduct.SSEStream(r.Context(), w, ch)
}

func handleEcho(w http.ResponseWriter, r *http.Request) {
	var req EchoRequest
	req.ID = chi.URLParam(r, "id")
	c, err := websocket.Accept(w, r, nil)
	if err != nil {
		goduct.WriteError(w, goduct.BadRequest("websocket accept failed"))
		return
	}
	defer c.CloseNow()
	_ = Echo(r.Context(), req, goduct.NewWSConn[EchoEvent, EchoMessage](c))
}