# examples/chi-basic/testdata/expected/echo/goduct_routes.go
// Code generated by goduct. DO NOT EDIT.

package api

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/coder/websocket"
	"github.com/labstack/echo/v4"

	goduct "github.com/townsendmerino/goduct/runtime"
)

// Register wires every goduct-annotated handler in this package to r.
// Call this after applying any middleware you want to share across handlers.
func Register(r *echo.Echo) {
	r.GET("/users/:id", handleGetUser)
	r.GET("/users", handleListUsers)
	r.POST("/users", handleCreateUser)
	r.PATCH("/users/:id", handleUpdateUser)
	r.DELETE("/users/:id", handleDeleteUser)
	r.POST("/users/:id/avatar", handleUploadAvatar)
	r.GET("/users/:id/events", handleWatchUserEvents)
	r.GET("/users/:id/echo", handleEcho)
}

func handleGetUser(c echo.Context) error {
	var req GetUserRequest
	req.ID = c.Param("id")
	q := c.Request().URL.Query()
	req.Include = q.Get("include")
	resp, err := GetUser(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	goduct.WriteJSON(c.Response().Writer, http.StatusOK, resp)
	return nil
}

func handleListUsers(c echo.Context) error {
	var req ListUsersRequest
	q := c.Request().URL.Query()
	if v := q.Get("limit"); v != "" {
		n, err := strconv.Atoi(v)
		if err != nil {
			goduct.WriteError(c.Response().Writer, goduct.BadRequest("limit must be an integer"))
			return nil
		}
		req.Limit = n
	}
	req.Cursor = q.Get("cursor")
	if v := q.Get("active"); v != "" {
		n, err := strconv.ParseBool(v)
		if err != nil {
			goduct.WriteError(c.Response().Writer, goduct.BadRequest("active must be a boolean"))
			return nil
		}
		req.Active = n
	}
	if v := q.Get("minScore"); v != "" {
		n, err := strconv.ParseFloat(v, 64)
		if err != nil {
			goduct.WriteError(c.Response().Writer, goduct.BadRequest("minScore must be a number"))
			return nil
		}
		req.MinScore = n
	}
	resp, err := ListUsers(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	goduct.WriteJSON(c.Response().Writer, http.StatusOK, resp)
	return nil
}

func handleCreateUser(c echo.Context) error {
	var req CreateUserRequest
	if err := json.NewDecoder(c.Request().Body).Decode(&req); err != nil {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("invalid json body"))
		return nil
	}
	resp, err := CreateUser(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	goduct.WriteJSON(c.Response().Writer, http.StatusCreated, resp)
	return nil
}

func handleUpdateUser(c echo.Context) error {
	var req UpdateUserRequest
	if err := json.NewDecoder(c.Request().Body).Decode(&req); err != nil {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("invalid json body"))
		return nil
	}
	req.ID = c.Param("id")
	resp, err := UpdateUser(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	goduct.WriteJSON(c.Response().Writer, http.StatusOK, resp)
	return nil
}

func handleDeleteUser(c echo.Context) error {
	var req DeleteUserRequest
	req.ID = c.Param("id")
	if err := DeleteUser(c.Request().Context(), req); err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	c.Response().Writer.WriteHeader(http.StatusNoContent)
	return nil
}

func handleUploadAvatar(c echo.Context) error {
	var req UploadAvatarRequest
	if err := c.Request().ParseMultipartForm(32 << 20); err != nil {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("invalid multipart form"))
		return nil
	}
	if files := c.Request().MultipartForm.File["file"]; len(files) > 0 {
		req.File = files[0]
	} else {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("file is required"))
		return nil
	}
	if req.File != nil && req.File.Size > 1048576 {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("file exceeds 1048576 byte limit"))
		return nil
	}
	if files := c.Request().MultipartForm.File["thumbnails"]; len(files) > 0 {
		req.Thumbnails = files
	}
	if vs := c.Request().MultipartForm.Value["caption"]; len(vs) > 0 {
		req.Caption = vs[0]
	}
	req.UserID = c.Param("id")
	resp, err := UploadAvatar(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	goduct.WriteJSON(c.Response().Writer, http.StatusCreated, resp)
	return nil
}

func handleWatchUserEvents(c echo.Context) error {
	var req WatchUserEventsRequest
	req.ID = c.Param("id")
	ch, err := WatchUserEvents(c.Request().Context(), req)
	if err != nil {
		goduct.WriteError(c.Response().Writer, err)
		return nil
	}
	c.Response().Writer.Header().Set("Content-Type", "text/event-stream")
	c.Response().Writer.Header().Set("Cache-Control", "no-cache")
	c.Response().Writer.WriteHeader(http.StatusOK)
	goduct.SSEStream(c.Request().Context(), c.Response().Writer, ch)
	return nil
}

func handleEcho(c echo.Context) error {
	var req EchoRequest
	req.ID = c.Param("id")
	c, err := websocket.Accept(c.Response().Writer, c.Request(), nil)
	if err != nil {
		goduct.WriteError(c.Response().Writer, goduct.BadRequest("websocket accept failed"))
		return nil
	}
	defer c.CloseNow()
	_ = Echo(c.Request().Context(), req, goduct.NewWSConn[EchoEvent, EchoMessage](c))
	return nil
}