# examples/chi-basic/testdata/expected/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "version": "0.0.0"
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Returns a page of users.",
        "operationId": "ListUsers",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minScore",
            "in": "query",
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Creates a new user.",
        "operationId": "CreateUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              },
              "example": {
                "email": "alice@example.com",
                "name": "Alice",
                "role": "member"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Returns a single user by ID.",
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "email": "alice@example.com",
                  "id": "u_1",
                  "name": "Alice",
                  "status": "active"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Removes a user.",
        "operationId": "DeleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Updates fields on an existing user.",
        "description": "Updates fields on an existing user. Omitted fields are not changed.",
        "operationId": "UpdateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/avatar": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Stores a new avatar image for the user.",
        "operationId": "UploadAvatar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "thumbnails": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/events": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Streams events for one user as they happen.",
        "operationId": "WatchUserEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream of UserEvent",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/UserEvent"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoductError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUserRequest": {
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "viewer",
              "member"
            ],
            "type": "string"
          },
          "website": {
            "format": "uri",
            "type": "string"
          },
          "referralCode": {
            "maxLength": 8,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "role"
        ],
        "type": "object"
      },
      "EchoEvent": {
        "properties": {
          "echo": {
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "at",
          "echo"
        ],
        "type": "object"
      },
      "EchoMessage": {
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
      "GoductError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ListUsersResponse": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "users"
        ],
        "type": "object"
      },
      "Profile": {
        "properties": {
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "bio",
          "tags"
        ],
        "type": "object"
      },
      "UpdateUserRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          }
        },
        "type": "object"
      },
      "UploadAvatarRequest": {
        "properties": {},
        "type": "object"
      },
      "User": {
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "email",
          "id",
          "name",
          "status"
        ],
        "type": "object"
      },
      "UserEvent": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "action",
          "at",
          "userId"
        ],
        "type": "object"
      },
      "UserStatus": {
        "enum": [
          "active",
          "invited",
          "suspended"
        ],
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "errors",
          "field"
        ],
        "type": "object"
      }
    }
  }
}